requirements.txt
requests
python-dotenv
tqdm
pandas
streamlit
fastapi
uvicorn
gunicorn
